requirements.txt
click==7.1.2
dnspython==2.1.0
Flask==1.1.2
Flask-PyMongo==2.3.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pymongo==3.11.2
python-dotenv==0.15.0
style==1.1.0
update==0.0.1
Werkzeug==1.0.1
